+++ openwhisk/compiler.go
@@ -54,15 +54,18 @@ func (ap *ActionProxy) CompileAction(main string, srcDir string, binDir string) 
 	var cmd *exec.Cmd
 	cmd = exec.Command(ap.compiler, main, srcDir, binDir)
 	cmd.Env = []string{"PATH=" + os.Getenv("PATH")}
+	for k, v := range ap.env {
+		cmd.Env = append(cmd.Env, k+"="+v)
+	}
 
 	// gather stdout and stderr
 	out, err := cmd.CombinedOutput()
 	Debug("compiler out: %s, %v", out, err)
-	if err != nil {
-		return err
-	}
 	if len(out) > 0 {
 		return fmt.Errorf("%s", out)
+	}
+	if err != nil {
+		return err
 	}
 	return nil
 }